+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -3,6 +3,7 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import web.org.springframework.web.DispatcherServlet;
 import web.org.springframework.web.WebApplicationInitializer;
 
 /**
